agent/requirements.txt
python-event-bus==1.2.0
PyYAML==6.0.1
setuptools==70.0.0
websockets==12.0
